requirements.txt
python-dotenv==0.21.1
python-telegram-bot==13.0
redis==4.5.1
vk-api==11.9.9
